+++ mods/mobs_redo/mobs_ghost_redo/init.lua
@@ -21,16 +21,9 @@
 
 --]]
 
---
--- localization
---
+-- Translation and settings
 
 local S = minetest.get_translator("mobs_ghost_redo")
-
---
--- General variables
---
-
 local ghost_daytime_check = minetest.settings:get_bool("mobs_ghost_redo_daytime_check")
 local ghost_bones_only = minetest.settings:get_bool("mobs_ghost_redo_bones_only")
 local ghost_difficulty = minetest.settings:get_bool("mobs_ghost_redo_difficulty")
@@ -47,15 +40,19 @@ if ghost_difficulty == nil then
 	ghost_difficulty = false
 end
 
+-- Mineclone check
+
+local mod_mcl = minetest.get_modpath("mcl_core")
+
+-- Spawn settings (default, bones only, mineclone)
+
 local SPAWNING_CHANCE = 7500
 local ACTIVE_OBJECTS = 2
-local SPAWNING_NODES = {
-	"group:cracky",
-	"group:stone",
-	"group:crumbly",
-	"group:sand",
-	"group:snowy",
-}
+local SPAWNING_NODES = {"group:cracky", "group:crumbly"}
+
+if mod_mcl == true then
+	SPAWNING_NODES = {"group:pickaxey", "group:shovely"}
+end
 
 if ghost_bones_only == true then
 	SPAWNING_NODES = {"bones:bones", "mobs_humans:human_bones"}
@@ -63,21 +60,14 @@ if ghost_bones_only == true then
 	ACTIVE_OBJECTS = 1
 end
 
---
 -- Functions
---
 
 local function is_daytime()
 
 	local time = minetest.get_timeofday() * 24000
 
-	if time >= 4700 and time <= 19250 then
-		return true
-	end
-
-	return false
+	if time >= 4700 and time <= 19250 then return true end
 end
-
 
 local function random_mesh()
 
@@ -88,9 +78,12 @@ local function random_mesh()
 	return "mobs_ghost_redo_ghost_1.b3d"
 end
 
---
+-- Drops
+
+local DROPS = {{name = (mod_mcl and "mcl_raw_ores:raw_gold" or "default:gold_lump"),
+		chance = 100, min = 1, max = 5}}
+
 -- Entity definition
---
 
 mobs:register_mob("mobs_ghost_redo:ghost", {
 	type = "monster",
@@ -120,34 +113,19 @@ mobs:register_mob("mobs_ghost_redo:ghost", {
 		damage = "mobs_ghost_redo_ghost_hit",
 		death = "mobs_ghost_redo_ghost_death"
 	},
-	drops = {
-		{name = "default:gold_lump", chance = 100, min = 1, max = 5}
-	},
+	drops = DROPS,
 	visual = "mesh",
 	visual_size = {x = 1, y = 1},
 	collisionbox = {-0.3, -0.5, -0.3, 0.3, 1.5, 0.3},
 	textures = {"mobs_ghost_redo_ghost.png"},
 	mesh = "mobs_ghost_redo_ghost_1.b3d",
 	animation = {
-		stand_start = 0,
-		stand_end = 80,
-		stand_speed = 15,
-		walk_start = 102,
-		walk_end = 122,
-		walk_speed = 12,
-		run_start = 102,
-		run_end = 122,
-		run_speed = 10,
-		fly_start = 102,
-		fly_end = 122,
-		fly_speed = 12,
-		punch_start = 102,
-		punch_end = 122,
-		punch_speed = 25,
-		die_start = 81,
-		die_end = 101,
-		die_speed = 28,
-		die_loop = false,
+		stand_start = 0, stand_end = 80, stand_speed = 15,
+		walk_start = 102, walk_end = 122, walk_speed = 12,
+		run_start = 102, run_end = 122, run_speed = 10,
+		fly_start = 102, fly_end = 122, fly_speed = 12,
+		punch_start = 102, punch_end = 122, punch_speed = 25,
+		die_start = 81, die_end = 101, die_speed = 28, die_loop = false
 	},
 
 	on_spawn = function(self, pos)
@@ -157,11 +135,11 @@ mobs:register_mob("mobs_ghost_redo:ghost", {
 			self.health = math.random(20, 30)
 
 			self.immune_to = {
-				{"default:sword_steel", 6},
+				{(mod_mcl and "mcl_tools:sword_iron" or "default:sword_steel"), 6},
 				{"default:sword_bronze", 6},
-				{"default:sword_mese", 7},
+				{(mod_mcl and "mcl_tools:sword_gold" or "default:sword_mese"), 7},
 				{"mobs_others:sword_obsidian", 7},
-				{"default:sword_diamond", 8},
+				{(mod_mcl and "mcl_tools:sword_diamond" or "default:sword_diamond"), 8},
 				{"moreores:sword_silver", 12},
 				{"moreores:sword_mithril", 9},
 				{"pigiron:sword_iron", 6},
@@ -242,41 +220,34 @@ mobs:register_mob("mobs_ghost_redo:ghost", {
 	end
 })
 
---
--- Ghost's spawn
---
+-- Ghost spawn - Check for custom spawn.lua
 
-mobs:spawn({name = "mobs_ghost_redo:ghost",
-	nodes = SPAWNING_NODES,
-	neighbors = {"air"},
-	max_light = 4,
-	interval = 60,
-	chance = SPAWNING_CHANCE,
-	active_object_count = ACTIVE_OBJECTS,
-	min_height = -30912,
-	day_toggle = false
-})
+local MP = minetest.get_modpath(minetest.get_current_modname()) .. "/"
+local input = io.open(MP .. "spawn.lua", "r")
 
---
+if input then
+	input:close() ; input = nil ; dofile(MP .. "spawn.lua")
+else
+	mobs:spawn({name = "mobs_ghost_redo:ghost",
+		nodes = SPAWNING_NODES,
+		neighbors = {"air"},
+		max_light = 4,
+		interval = 60,
+		chance = SPAWNING_CHANCE,
+		active_object_count = ACTIVE_OBJECTS,
+		min_height = -30912,
+		day_toggle = false
+	})
+end
+
 -- Ghost's egg
---
 
 mobs:register_egg("mobs_ghost_redo:ghost", S("Ghost Spawner"),
 	"mobs_ghost_redo_egg_ghost.png", 0)
 
---
 -- Alias
---
 
 mobs:alias_mob("mobs:ghost", "mobs_ghost_redo:ghost")
 
---
--- Minetest engine debug logging
---
 
-if minetest.settings:get("debug_log_level") == nil
-or minetest.settings:get("debug_log_level") == "action"
-or minetest.settings:get("debug_log_level") == "info"
-or minetest.settings:get("debug_log_level") == "verbose" then
-	minetest.log("action", "[Mod] Mobs Ghost Redo [v0.7.0] loaded.")
-end
+print("[MOD] Mobs Ghost Redo loaded.")